pyproject.toml
[project]
name = "pg-vla"
version = "0.1.0"
description = "Training vision-language-action models"
readme = "README.md"
requires-python = "==3.10.*"
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "einops",
    "gsutil>=5.32",
    "hydra-core",
    "matplotlib",
    "numpy",
    "omegaconf",
    "pillow",
    "pre-commit>=4.0.1",
    "pretty_errors",
    "protobuf==3.20.3",
    "tensorflow==2.15.0",
    "tensorflow_datasets==4.9.2",
    "torch==2.5.0",
    "transformers",
    "tqdm",
    "wandb",
]

[project.urls]
Homepage = "https://pg-vla.github.io/"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = []

[tool.ruff]
line-length = 88
target-version = "py310"
extend-exclude = ["src/data/obs_transforms.py", "src/data/utils/data_utils.py"]

[tool.ruff.lint]
select = ["A", "B", "E", "F", "I", "RUF", "W"]
ignore = ["E203", "E501", "B006"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403"]
